fix: BacktestBroker preserves fractional qty for bracket orders (#14)
Issue #9 surfaced after PR #6 wired live MomentumStrategy through
BacktestEngine: bracket orders flowed to the broker but were logged
with quantity=0 and produced flat equity curves.

Initial diagnosis (now revised): the issue title attributed the bug to
qty extraction failing on bracket envelopes — i.e., that qty lived
inside a leg structure and getattr(order_request, "qty") was returning
None. Empirically that's not what alpaca-py's MarketOrderRequest looks
like for brackets: qty is a top-level float, no leg wrapping. The
getattr extraction worked correctly all along.

Actual cause: submit_order_advanced extracted qty as a float (e.g.,
0.4996), then truncated it to int(qty) before calling place_order.
Sub-1 quantities collapsed to 0; fractional ≥1 quantities silently
shrank (e.g., 4.95 → 4). Cast to float instead — place_order's
arithmetic and trade record were already float-safe.

AlpacaBroker.submit_order_advanced does not have this bug — it passes
the order_request through to alpaca-py's submit_order unchanged, so
fractional shares hit the live API correctly. No live regression risk.

Tests:
- Strengthen tests/unit/test_momentum_strategy_backtest_parity.py
  to assert non-zero quantity in the recorded trade. Pre-fix this
  assertion would fail with quantity=0; post-fix records 0.4996.
- Add tests/unit/test_backtest_broker.py
  test_submit_order_advanced_preserves_fractional_qty_for_bracket
  exercising real OrderBuilder→MarketOrderRequest envelopes for
  both sub-1 (0.5) and fractional ≥1 (4.95) cases, end-to-end through
  the broker including position book and trade record.

Closes #9

+++ brokers/backtest_broker.py
@@ -759,8 +759,11 @@ class BacktestBroker:
         if hasattr(side, "value"):
             side = side.value
 
-        # Place the order
-        result = self.place_order(symbol, int(qty), side, order_type=str(order_type))
+        # Place the order. Cast to float (not int) so fractional-share orders
+        # — produced legitimately by Kelly/volatility sizing on small accounts
+        # and by Alpaca paper trading's native fractional-share support —
+        # don't get truncated to 0 (sub-1 qty) or silently shrunk (e.g., 4.95→4).
+        result = self.place_order(symbol, float(qty), side, order_type=str(order_type))
 
         # Return a mock order response
         class MockOrder:
